test(vm): move last start time tests to statistics
